refact(insights): move more code to common
