+++ instantsearch-insights/src/androidMain/kotlin/com/algolia/instantsearch/insights/internal/extension/Time.kt
@@ -1,4 +1,4 @@
 package com.algolia.instantsearch.insights.internal.extension
 
-internal val currentTimeMillis: Long
+internal actual val currentTimeMillis: Long
     get() = System.currentTimeMillis()
