+++ instantsearch-insights/src/commonMain/kotlin/com/algolia/instantsearch/insights/FilterTrackable.kt
@@ -1,6 +1,6 @@
 package com.algolia.instantsearch.insights
 
-import com.algolia.instantsearch.insights.internal.currentTimeMillis
+import com.algolia.instantsearch.insights.internal.extension.currentTimeMillis
 import com.algolia.search.model.filter.Filter
 import com.algolia.search.model.insights.EventName
