--- instantsearch-insights/src/commonMain/kotlin/com/algolia/instantsearch/insights/internal/Time.kt
@@ -1,3 +1,0 @@
-package com.algolia.instantsearch.insights.internal
-
-internal expect val currentTimeMillis: Long
