+++ instantsearch-insights/src/commonMain/kotlin/com/algolia/instantsearch/insights/internal/cache/InsightsEventCache.kt
