+++ instantsearch-insights/src/commonMain/kotlin/com/algolia/instantsearch/insights/internal/data/local/mapper/InsightsEventsMapper.kt
