+++ instantsearch-insights/src/commonMain/kotlin/com/algolia/instantsearch/insights/internal/data/settings/InsightsSettings.kt
