+++ instantsearch-insights/src/commonMain/kotlin/com/algolia/instantsearch/insights/internal/event/EventResponse.kt
