+++ instantsearch-insights/src/commonMain/kotlin/com/algolia/instantsearch/insights/internal/extension/Map.kt
