+++ instantsearch-insights/src/commonMain/kotlin/com/algolia/instantsearch/insights/internal/extension/Time.kt
@@ -1,0 +1,3 @@
+package com.algolia.instantsearch.insights.internal.extension
+
+internal expect val currentTimeMillis: Long
