+++ instantsearch-insights/src/commonMain/kotlin/com/algolia/instantsearch/insights/internal/logging/InsightsLogger.kt
@@ -1,0 +1,12 @@
+package com.algolia.instantsearch.insights.internal.logging
+
+import com.algolia.search.model.IndexName
+
+internal expect object InsightsLogger {
+
+    var enabled: MutableMap<IndexName, Boolean>
+
+    fun log(indexName: IndexName, message: String)
+
+    fun log(message: String)
+}
