+++ instantsearch-insights/src/commonMain/kotlin/com/algolia/instantsearch/insights/internal/uploader/InsightsUploader.kt
